refactor: change folder and file names with linting
